requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.0.2
joblib==1.4.2
pydantic==2.10.3
python-multipart==0.0.20

# Database dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
